test/requirements.txt
sorl-thumbnail
easy-thumbnails
pillow
pytest
pytest-django
pytest-pythonpath
pytest-cov
pytest-xdist
pytest-forked
